vmwos: atags: fix cmdline dump

+++ kernel/string.h
@@ -1,6 +1,6 @@
 int strncmp(const char *s1, const char *s2, uint32_t n);
 void *memcpy(void *dest, const void *src, uint32_t n);
-char *strncpy(char *dest, const char *src, size_t n);
+char *strncpy(char *dest, const char *src, uint32_t n);
 int memcmp(const char *s1, const char *s2, uint32_t n);
 int strlen(const char *s1);
 void *memset(void *s, int c, uint32_t n);
